move pytest_addoption to the root
