+++ tests/charm/integration/conftest.py
@@ -9,21 +9,9 @@ import pytest
 import pytest_asyncio
 from juju.application import Application
 from juju.model import Model
-from pytest import Parser
 from pytest_operator.plugin import OpsTest
 
-CHARM_FILE_PARAM = "--charm-file"
-FLASK_APP_IMAGE_PARAM = "--repo-policy-compliance-image"
-
-
-def pytest_addoption(parser: Parser) -> None:
-    """Parse additional pytest options.
-
-    Args:
-        parser: Pytest parser.
-    """
-    parser.addoption(CHARM_FILE_PARAM, action="store", help="Charm file to be deployed")
-    parser.addoption(FLASK_APP_IMAGE_PARAM, action="store", help="Flask app image to be deployed")
+from tests.conftest import CHARM_FILE_PARAM, FLASK_APP_IMAGE_PARAM
 
 
 @pytest.fixture(scope="session", name="github_token")
